% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueRsky-package.R
\docType{package}
\name{blueRsky-package}
\alias{blueRsky}
\alias{blueRsky-package}
\title{blueRsky: Talk to bluesky}
\description{
This package allows you to talk to bluesky from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/blueRsky}
  \item Report bugs at \url{https://github.com/colin-fraser/blueRsky/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
