% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_create_post}
\alias{bsky_create_post}
\title{Create a new post}
\usage{
bsky_create_post(text, user_did = blueRsky::get_session_did())
}
\arguments{
\item{text}{A character string containing the text of the post.}

\item{user_did}{An at-identifier representing the user's DID. Defaults to the current session DID.}
}
\value{
A list containing the created post's information.
}
\description{
This function creates a new post in the user's feed.
}
\examples{
\dontrun{
  bsky_login("username", "password")
  result <- bsky_create_post("Hello, world!")
  print(result)
}
}
