% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_delete_post}
\alias{bsky_delete_post}
\title{Delete a post}
\usage{
bsky_delete_post(post_key, user_did = blueRsky::get_session_did())
}
\arguments{
\item{post_key}{A character string representing the post's key.}

\item{user_did}{An at-identifier representing the user's DID. Defaults to the current session DID.}
}
\value{
An HTTP response object. If successful, the status code will be 200.
}
\description{
This function deletes a post from the user's feed.
}
\examples{
\dontrun{
  bsky_login("username", "password")
  post_key <- "example_post_key"
  response <- bsky_delete_post(post_key)
  print(response$status_code)
}
}
