% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_describe_repo}
\alias{bsky_describe_repo}
\title{Get information about the repo, including the list of collections.}
\usage{
bsky_describe_repo(
  repo = blueRsky::get_session_handle(),
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{repo}{at-identifier}

\item{...}{Additional arguments to requestor}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
\link{Describe the return value here}
}
\description{
Get information about the repo, including the list of collections.
}
