% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_followers}
\alias{bsky_get_followers}
\title{Who is following an actor?}
\usage{
bsky_get_followers(
  actor = blueRsky::get_session_handle(),
  limit = NULL,
  cursor = NULL,
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{actor}{at-identifier}

\item{limit}{number between 1 and 100}

\item{cursor}{cursor}

\item{...}{Additional arguments to requestor}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
list of followers
}
\description{
Who is following an actor?
}
