% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_profile}
\alias{bsky_get_profile}
\title{Get Profile}
\usage{
bsky_get_profile(
  actor = blueRsky::get_session_handle(),
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{actor}{Description of actor. Defaults to the logged-in user.}

\item{...}{Additional arguments passed to requestor}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
Profile
}
\description{
Get Profile
}
