% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_session}
\alias{bsky_get_session}
\title{Get information about the current session.}
\usage{
bsky_get_session(...)
}
\arguments{
\item{...}{Additional arguments to requestor}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
\link{Describe the return value here}
}
\description{
Get information about the current session.
}
