% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_profile}
\alias{bsky_get_profile}
\title{Get Profile}
\usage{
bsky_get_profile(...)
}
\arguments{
\item{...}{Additional arguments passed to requestor}

\item{actor}{Description of actor. Defaults to the logged-in user.}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
Profile
}
\description{
Get Profile
}
