% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_profile}
\alias{bsky_get_profile}
\title{Get Profile}
\usage{
bsky_get_profile(
  actor = blueRsky::get_session_handle(),
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{actor}{Description of actor}

\item{...}{Description of ...}

\item{credentials}{Description of credentials}

\item{action}{Description of action}

\item{decode_if_success}{Description of decode_if_success}
}
\value{
\link{Describe the return value here}
}
\description{
Get Profile
}
