% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluesky.R
\name{bsky_get_timeline}
\alias{bsky_get_timeline}
\title{A view of the user's home timeline.}
\usage{
bsky_get_timeline(
  algorithm = NULL,
  limit = NULL,
  cursor = NULL,
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{algorithm}{Algorithm to user}

\item{limit}{number between 1 and 100}

\item{cursor}{cursor}

\item{...}{Additional arguments passed to requestor}

\item{credentials}{Credential.}

\item{action}{Either 'perform' or 'dry-run'}

\item{decode_if_success}{If FALSE, returns the whole response object. If true, returns the decoded response.}
}
\value{
A list of home timeline objects
}
\description{
A view of the user's home timeline.
}
